Move resolution tests to separate file
